[Sprint 2][Code Refactor]
