+++ backend/go.mod
@@ -1,5 +1,5 @@
-module github.com/richagupta04
-
+//module github.com/richagupta04
+module github.com/anjaligupta0621/EasyConnect/backend
 go 1.17
 
 require github.com/jinzhu/gorm v1.9.16
