+++ backend/handler/endpoints.go
@@ -1,28 +1,12 @@
-package main
+package handler
 
 import (
 	"encoding/json"
 	"net/http"
 
+	"github.com/anjaligupta0621/EasyConnect/backend/models"
 	"github.com/jinzhu/gorm"
 )
-
-// Our User Struct
-type Recruiter struct {
-	ID           uint   `gorm:"primary_key; AUTO_ Increment "`
-	Name         string `gorm:"Not null "`
-	Email        string `gorm:"unique; Not null "`
-	Password     string `gorm:"unique; Not null "`
-	Organization string
-	Website      string
-	Contact      string `gorm:"unique; Not null "`
-	Jobs         []Job  `gorm:"foreignKey:JobID;constraint:OnUpdate:CASCADE,OnDelete:RESTRICT;"`
-}
-
-type Login struct {
-	Email    string
-	Password string
-}
 
 func setupCorsResponse(w *http.ResponseWriter, req *http.Request) {
 	(*w).Header().Set("Access-Control-Allow-Origin", "*")
@@ -39,13 +23,13 @@ func getUsers(w http.ResponseWriter, r *http.Request) {
 	defer db.Close()
 
 	decoder := json.NewDecoder(r.Body)
-	var login Login
+	var login models.Login
 	err2 := decoder.Decode(&login)
 	if err2 != nil {
 		panic(err2)
 	}
 
-	var recruiter_ Recruiter
+	var recruiter_ models.Recruiter
 	db.Table("recruiters").Where("Email = ? AND Password = ?", login.Email, login.Password).Find(&recruiter_)
 
 	if recruiter_.Email != "" {
@@ -64,12 +48,12 @@ func putUserData(w http.ResponseWriter, r *http.Request) {
 	defer db.Close()
 
 	decoder := json.NewDecoder(r.Body)
-	var recruiter Recruiter
+	var recruiter models.Recruiter
 	err2 := decoder.Decode(&recruiter)
 	if err2 != nil {
 		panic(err2)
 	}
 
-	db.Create(&Recruiter{Name: recruiter.Name, Email: recruiter.Email, Password: recruiter.Password, Organization: recruiter.Organization, Website: recruiter.Website, Contact: recruiter.Contact})
+	db.Create(&models.Recruiter{Name: recruiter.Name, Email: recruiter.Email, Password: recruiter.Password, Organization: recruiter.Organization, Website: recruiter.Website, Contact: recruiter.Contact})
 	json.NewEncoder(w).Encode("New Recruiter Successfully Added: " + recruiter.Name)
 }
